--- packages/blocks/src/extensions/index.ts
@@ -1,3 +1,0 @@
-import { main as ledits } from './ledits/main.js';
-
-export const extensions = [ledits];
